# target/classes/openapi.yaml
openapi: 3.0.3
info:
  title: Klassenplaner API
  version: 1.0.0

servers:
  - url: /api/v1

paths:

  /auth/login:
    post:
      summary: Einloggen
      requestBody:
        content:
          application/json:
            schema:
              properties:
                username: { type: string, example: "AP26a" }
                password: { type: string, example: "geheim" }
      responses:
        '200':
          description: Login erfolgreich – JWT zurückgegeben
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Benutzername oder Passwort falsch

  /auth/logout:
    post:
      summary: Ausloggen
      responses:
        '204':
          description: Logout erfolgreich

  /auth/register:
    post:
      summary: Neue Klasse registrieren (optional)
      requestBody:
        content:
          application/json:
            schema:
              properties:
                className: { type: string, example: "AP26a" }
                password:  { type: string, example: "geheim" }
      responses:
        '201':
          description: Klasse wurde angelegt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Klasse'
        '409':
          description: Klasse existiert bereits

  /classes/{classId}/modules:
    get:
      summary: Alle Module abrufen
      parameters:
        - { name: classId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Liste aller Module der Klasse
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Modul'
    post:
      summary: Neues Modul anlegen
      parameters:
        - { name: classId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name: { type: string, example: "Mathematik" }
      responses:
        '201':
          description: Modul wurde angelegt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Modul'
        '400':
          description: Name darf nicht leer sein

  /modules/{moduleId}/homework:
    get:
      summary: Alle Hausaufgaben abrufen (neueste zuerst)
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Liste aller Hausaufgaben
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hausaufgabe'
    post:
      summary: Neue Hausaufgabe erfassen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:       { type: string, example: "Aufgaben S. 42" }
                description: { type: string, example: "Aufgaben 1–10 lösen" }
                dueDate:     { type: string, format: date, example: '2026-06-10' }
      responses:
        '201':
          description: Hausaufgabe wurde gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hausaufgabe'
        '400':
          description: Titel oder Datum fehlt

  /modules/{moduleId}/homework/{id}:
    delete:
      summary: Hausaufgabe löschen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
        - { name: id,       in: path, required: true, schema: { type: integer } }
      responses:
        '204':
          description: Hausaufgabe wurde gelöscht
        '404':
          description: Hausaufgabe nicht gefunden

  /modules/{moduleId}/exams:
    get:
      summary: Alle Prüfungen abrufen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Liste aller Prüfungstermine
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pruefung'
    post:
      summary: Neuen Prüfungstermin erfassen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              properties:
                topic:         { type: string, example: "Integralrechnung" }
                examDate:      { type: string, format: date, example: '2026-06-20' }
                learningGoals: { type: string, example: "Bestimmtes Integral berechnen" }
      responses:
        '201':
          description: Prüfungstermin wurde gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pruefung'
        '400':
          description: Thema oder Datum fehlt

  /modules/{moduleId}/exams/{id}:
    delete:
      summary: Prüfungstermin löschen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
        - { name: id,       in: path, required: true, schema: { type: integer } }
      responses:
        '204':
          description: Prüfungstermin wurde gelöscht
        '404':
          description: Prüfungstermin nicht gefunden

  /modules/{moduleId}/infos:
    get:
      summary: Alle Informationen abrufen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Liste aller Informationseinträge
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Information'
    post:
      summary: Neuen Informationseintrag erfassen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              properties:
                title:   { type: string, example: "Raumänderung" }
                content: { type: string, example: "Ab nächster Woche Zimmer 204" }
      responses:
        '201':
          description: Informationseintrag wurde gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Information'
        '400':
          description: Titel oder Text fehlt

  /modules/{moduleId}/infos/{id}:
    delete:
      summary: Informationseintrag löschen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
        - { name: id,       in: path, required: true, schema: { type: integer } }
      responses:
        '204':
          description: Informationseintrag wurde gelöscht
        '404':
          description: Eintrag nicht gefunden

  /modules/{moduleId}/questions:
    get:
      summary: Alle Fragen abrufen (inkl. Antworten)
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Liste aller Fragen mit ihren Antworten
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Frage'
    post:
      summary: Neue Frage stellen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              properties:
                questionText: { type: string, example: "Was kommt an der Prüfung dran?" }
      responses:
        '201':
          description: Frage wurde gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Frage'
        '400':
          description: Fragetext darf nicht leer sein

  /modules/{moduleId}/questions/{id}:
    delete:
      summary: Frage löschen (inkl. alle Antworten)
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
        - { name: id,       in: path, required: true, schema: { type: integer } }
      responses:
        '204':
          description: Frage und Antworten wurden gelöscht
        '404':
          description: Frage nicht gefunden

  /modules/{moduleId}/questions/{id}/answers:
    post:
      summary: Antwort auf eine Frage hinzufügen
      parameters:
        - { name: moduleId, in: path, required: true, schema: { type: integer } }
        - { name: id,       in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              properties:
                answerText: { type: string, example: "Kapitel 3–5, Schwerpunkt Ableitungen" }
      responses:
        '201':
          description: Antwort wurde gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Antwort'
        '404':
          description: Frage nicht gefunden

  /health:
    get:
      summary: Server-Status prüfen
      responses:
        '200':
          description: Server läuft

components:
  schemas:

    Token:
      type: object
      properties:
        token: { type: string, example: "eyJhbGciOiJIUzI1NiJ9..." }

    Klasse:
      type: object
      properties:
        id:        { type: integer, example: "1" }
        className: { type: string, example: "AP26a" }

    Modul:
      type: object
      properties:
        id:      { type: integer, example: "1" }
        name:    { type: string, example: "Mathematik" }
        classId: { type: integer, example: "1" }

    Hausaufgabe:
      type: object
      properties:
        id:          { type: integer, example: "1" }
        title:       { type: string, example: "Aufgaben S. 42" }
        description: { type: string, example: "Aufgaben 1–10 lösen" }
        dueDate:     { type: string, format: date, example: '2026-06-10' }
        createdAt:   { type: string, format: date-time }

    Pruefung:
      type: object
      properties:
        id:            { type: integer, example: "1" }
        topic:         { type: string, example: "Integralrechnung" }
        examDate:      { type: string, format: date, example: '2026-06-20' }
        learningGoals: { type: string, example: "Bestimmtes Integral berechnen" }
        createdAt:     { type: string, format: date-time }

    Information:
      type: object
      properties:
        id:        { type: integer, example: "1" }
        title:     { type: string, example: "Raumänderung" }
        content:   { type: string, example: "Ab nächster Woche Zimmer 204" }
        createdAt: { type: string, format: date-time }

    Frage:
      type: object
      properties:
        id:           { type: integer, example: "1" }
        questionText: { type: string, example: "Was kommt an der Prüfung dran?" }
        createdAt:    { type: string, format: date-time }
        answers:
          type: array
          items:
            $ref: '#/components/schemas/Antwort'

    Antwort:
      type: object
      properties:
        id:         { type: integer, example: "1" }
        answerText: { type: string, example: "Kapitel 3–5, Schwerpunkt Ableitungen" }
        createdAt:  { type: string, format: date-time }